3_distributed_training/time-series-forecasting/amazon-chronos2/sagemaker_code/djl_inference/requirements.txt
chronos-forecasting[extras]>=2.2
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.30.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
boto3>=1.28.0
